add optional filter by date in exportVocabulary

+++ src/main/java/ovh/miroslaw/kindle2anki/vocabulary/model/Vocabulary.java
@@ -12,8 +12,9 @@ import lombok.Getter;
 @Entity
 @Table(name = "WORDS")
 public class Vocabulary {
+
     @Id
     private String id;
     private String stem;
-    private Integer timestamp;
+    private Long timestamp;
 }
